[feat] #1 매물 자료 상태 배너 분리
